lightweightdjango/requirements.txt
Django==1.7.5
Pillow==2.7.0
